change int to float in 04-average

+++ 01-arrays/mm04-average/mm04-average.ino
@@ -11,7 +11,7 @@ void setup() {
 
 void loop() {
   randomArray();
-  int avg = getAverage();
+  float avg = getAverage();
   printArray(avg);
   delay(5000);
 }
